Move development deps under proper category
Development dependencies should be declared under the
"devDependencies" section in package.json.

Beside being helpful to people who analyze the codebase, this
distinction is also useful to a lot of automated tools.

A development dependency should never be declared as being a
production dependency unless there is a compelling reason.