Remove custom lint rules
The state of linting was kind of a mess, as there are two
differing linting mechanisms: The first one being the built-in
linter of CRA, and the second one being a custom set of linting
rules defined in the .eslintrc.js file.

The first problem is that it is not clear which one is the main
linter because they both output a load of errors. Looks like
nobody cares much about either in any case.

And the second problem is that the idea of introducing an
alternative linting mechanism to CRA is a bad idea. The whole
point of CRA is to help devs avoid configuration by providing
opinionated defaults. Nonetheless if one absolutely has to
extend/replace the CRA linter config, they should aim for a
compatible configuration at the very least. That doesn't seem to
be the case here.

After examining the errors generated by the alternative linter
configuration, I conclude that they are not really of a big help
(if not a problem), so they should be cleaned up.

I suspect that a previous dev introduced this alternative config
to integrate the linter errors with their code editor but I don't
think this is worth the confusion & complication. The way of CRA
is to follow the linter errors from the command line.

But if one wants to have a linter - editor integration really
badly, they should do it themselves, at their own risk: Use a
custom lint configuration, but not commit it to the project. Not
everybody should be confused because of one person's editing
preferences.

References:

- https://create-react-app.dev/docs/setting-up-your-editor/#displaying-lint-output-in-the-editor
- https://www.npmjs.com/package/eslint-config-react-app