Make implicit dependencies explicit
There are a couple of external libraries that are directly
imported into the app, but not declared in package.json. They
don't cause errors as they are dependencies of other explicit
dependencies so they are already installed.

It's much better to have direct dependencies declared explicitly
to avoid potential versioning related problems which can be very
obscure at times.

In any case it's good to see the full list of dependencies in
package.json for multiple other purposes as well. e.g. analysis,
refactoring..

Missing deps detected by
[depcheck](https://www.npmjs.com/package/depcheck).