Refactor routes
The component that declares the routes is so complicated that it
is hard to understand what route points to what component quickly.

(Thank you React Router for enabling developers to create routes
as components resulting in a mess)

This is not acceptable. All the routes should be able to be seen
in a glance. There shouldn't be logic in routing components. There
shouldn't be hidden routes inside container components. Those
things make it hard to analyze and work with a project.

This commit takes care of all the mentioned issues and produces a
clean route module, where every route and mapped component can be
seen clearly in a glance.

Note that some components needed to be refactored as they were
tightly coupled to the routing logic. I've liberated them.

Another note is that I converted some Typescript files to
Javascript because I need to ship this quickly and don't have time
to deal with Typescript's peculiarities. It is doubtful that
Typescript is being used properly in this project anyway. In the
end, I know this is going to make the job of the reviewer harder,
but I'm sorry, we have to pragmatic.