Implement WalletConnect support
Since this WalletConnect implementation is a replacement for the
KeySafe Chrome extension, the WalletConnect client is designed to
be %100 API compatible with it. So there is no need to make
changes in any other parts of the app.

The only part of the app that is updated is the account dropdown
on the top right corner, where a "logout" button is added. There
was arguably no need for it before, as the users could easily log
out from the Keysafe as well.