Refactor Review component and remove unnecessary code
Remove the line that was previously used for testing purposes in the Review component of the CreateEntityFlow. This line was not needed and has been removed to improve code cleanliness.

Fix console log in useCreateEntity hook

Remove a console log statement that was accidentally left in the useCreateEntity hook. This console log was not necessary and has been removed to clean up the code.

Update accordedRight in EditProperty component

Update the accordedRight property in the EditProperty component of the CurrentEntity Dashboard. The property is now assigned the value of editEntity.accordedRight if it is an array, otherwise it is left undefined. This change improves the handling of the accordedRight property in the component.